Change utils placement
