+++ django_prices/forms.py
@@ -7,7 +7,7 @@ from django import forms
 from django.core.validators import ValidationError
 from prices import Money
 
-from .locale_utils import get_locale_data
+from .utils.locale import get_locale_data
 from .validators import MaxMoneyValidator, MinMoneyValidator, MoneyPrecisionValidator
 from .widgets import FixedCurrencyMoneyInput, MoneyInput
